In process of resolving lint errors, currently resolved 400 errors and ignored files and lines accordingly.
