+++ public/language/en-GB/topic.json
@@ -7,7 +7,6 @@
 
 	"post_is_deleted": "This post is deleted!",
 	"topic_is_deleted": "This topic is deleted!",
-
 	"profile": "Profile",
 	"posted_by": "Posted by %1",
 	"posted_by_guest": "Posted by Guest",
