+++ src/api/users.js
@@ -474,12 +474,5 @@ usersAPI.generateExport = async (caller, { uid, type }) => {
             targetUid: uid,
             ip: caller.ip,
         });
-        
     });
-    
-
 };
-
-
-
-
