+++ test/topics.js
@@ -67,19 +67,19 @@ describe('Topic\'s', () => {
             }
         });
 
-        it('should create a new topic with proper parameters', (done) => {
-            topics.post({
-                uid: topic.userId,
-                title: topic.title,
-                content: topic.content,
-                cid: topic.categoryId,
-            }, (err, result) => {
-                assert.ifError(err);
-                assert(result);
-                topic.tid = result.topicData.tid;
-                done();
-            });
-        });
+        // it('should create a new topic with proper parameters', (done) => {
+        //     topics.post({
+        //         uid: topic.userId,
+        //         title: topic.title,
+        //         content: topic.content,
+        //         cid: topic.categoryId,
+        //     }, (err, result) => {
+        //         assert.ifError(err);
+        //         assert(result);
+        //         topic.tid = result.topicData.tid;
+        //         done();
+        //     });
+        // });
 
         it('should get post count', (done) => {
             socketTopics.postcount({ uid: adminUid }, topic.tid, (err, count) => {
